Modified lockfile checking method
Previous check looked to see if there was a lockfile, then if the PID in
the lockfile referred to a running process.  This created a false
positive if, after a reboot, a different process had gotten the PID in
the lockfile.  The new check looks to see if there's a running instance
of btsyncindicator with the PID in the lockfile.